+++ src/desugar/forloops.c
@@ -73,16 +73,16 @@ node_st *DFLstmts(node_st *node) {
             node_st *positive_step_overflow = NULL;
             node_st *positive_step_no_overflow = NULL;
 
-            node_st *positive_step_clamp_assign =
-                ASTassign(CCNcopy(clamped_end_ref),
-                          ASTbinop(ASTint(INT_MIN), CCNcopy(step_ref), BO_sub));
+            node_st *positive_step_clamp_assign = ASTassign(
+                CCNcopy(clamped_end_ref),
+                ASTbinop(ASTint(INT64_MIN), CCNcopy(step_ref), BO_sub));
             VARREF_WRITE(ASSIGN_REF(positive_step_clamp_assign)) = true;
 
             if (!(NODE_TYPE(ASSIGN_EXPR(step_assign)) == NT_INT &&
                   (INT_VAL(ASSIGN_EXPR(step_assign)) == 1 ||
                    (NODE_TYPE(ASSIGN_EXPR(end_assign)) == NT_INT &&
                     INT_VAL(ASSIGN_EXPR(end_assign)) <=
-                        INT_MIN - INT_VAL(ASSIGN_EXPR(step_assign)))))) {
+                        INT64_MIN - INT_VAL(ASSIGN_EXPR(step_assign)))))) {
 
                 node_st *positive_step_overflow_while =
                     ASTwhile(ASTbinop(CCNcopy(start_ref),
@@ -102,7 +102,7 @@ node_st *DFLstmts(node_st *node) {
                   (INT_VAL(ASSIGN_EXPR(step_assign)) != 1 ||
                    (NODE_TYPE(ASSIGN_EXPR(end_assign)) == NT_INT &&
                     INT_VAL(ASSIGN_EXPR(end_assign))) >
-                       INT_MIN - INT_VAL(ASSIGN_EXPR(step_assign))))) {
+                       INT64_MIN - INT_VAL(ASSIGN_EXPR(step_assign))))) {
                 node_st *positive_step_no_overflow_while = ASTwhile(
                     ASTbinop(CCNcopy(start_ref), CCNcopy(end_ref), BO_lt),
                     CCNcopy(FOR_STMTS(stmt)));
@@ -124,16 +124,16 @@ node_st *DFLstmts(node_st *node) {
             node_st *negative_step_overflow = NULL;
             node_st *negative_step_no_overflow = NULL;
 
-            node_st *negative_step_clamp_assign =
-                ASTassign(CCNcopy(clamped_end_ref),
-                          ASTbinop(ASTint(INT_MAX), CCNcopy(step_ref), BO_sub));
+            node_st *negative_step_clamp_assign = ASTassign(
+                CCNcopy(clamped_end_ref),
+                ASTbinop(ASTint(INT64_MAX), CCNcopy(step_ref), BO_sub));
             VARREF_WRITE(ASSIGN_REF(negative_step_clamp_assign)) = true;
 
             if (!(NODE_TYPE(ASSIGN_EXPR(step_assign)) == NT_INT &&
                   (INT_VAL(ASSIGN_EXPR(step_assign)) == -1 ||
                    (NODE_TYPE(ASSIGN_EXPR(end_assign)) == NT_INT &&
                     INT_VAL(ASSIGN_EXPR(end_assign)) <=
-                        INT_MAX - INT_VAL(ASSIGN_EXPR(step_assign)))))) {
+                        INT64_MAX - INT_VAL(ASSIGN_EXPR(step_assign)))))) {
                 node_st *negative_step_overflow_while =
                     ASTwhile(ASTbinop(CCNcopy(start_ref),
                                       CCNcopy(clamped_end_ref), BO_gt),
@@ -152,7 +152,7 @@ node_st *DFLstmts(node_st *node) {
                   (INT_VAL(ASSIGN_EXPR(step_assign)) != -1 ||
                    (NODE_TYPE(ASSIGN_EXPR(end_assign)) == NT_INT &&
                     INT_VAL(ASSIGN_EXPR(end_assign)) >
-                        INT_MAX - INT_VAL(ASSIGN_EXPR(step_assign)))))) {
+                        INT64_MAX - INT_VAL(ASSIGN_EXPR(step_assign)))))) {
                 node_st *negative_step_no_overflow_while = ASTwhile(
                     ASTbinop(CCNcopy(start_ref), CCNcopy(end_ref), BO_gt),
                     CCNcopy(FOR_STMTS(stmt)));
